Protect memo_fetch (project level) with a lock
This both works around a `cachecontrol` issue and prevents a little bit
of duplicate work.  There is a small race condition described below:

As background, `_fetch_project` fetches the project page for a project,
and also queues up a fetch for the max version's metadata (since quite
often that's what we'll need).

The `_fetch_project_metadata` also queues up a project fetch for most
projects it depends on.

Both of those happen in a threadpool, and can run concurrently.

There is a case where the max version isn't the one we want, so if

```
A has version 1 that depends on B < 2

B has versions 1 and 2 that depend on C

C is a boring leaf
```

the prefetch of metadata is for B==2 but once the main thread gets B it
will realize that it needs to get metadata for B==1 after
`find_best_compatible_version` which is intentionally serial for
repeatable solves.

We might end up schedule both versions of B `_fetch_project_metadata`
about the same time if the main thread is actively waiting on B's
future, and that schedules C _twice_.  This lock ensure that we only
create one future for C, and shouldn't drastically affect performance
(ThreadPoolExecutor.submit is already protected with a lock.)

Refs #27