Fix a bug in --have handling
There was no test coverage for this case before, and I didn't notice
this in line coverage because it was an omitted `else: pass`.

Caused by the optimization in #26, the logic here was not correct, and
the version not being in `candidates` is actually the common case, so we
just check `requires_python` again.