Only log once when requirements.txt lines get ignored
We used to re-parse the file several times due to suggest mode, which
would print "Ignoring" on stdout several times (both above and below the
tree).  Cache the parsing, but also simplify the message to just give a
heads-up.  The lines themselves can still be shown with `--vmodule
hdeps.requirements=0` or above.

With a simple scenario test for ignored requirements too.

Fixes #32